requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audioop-lts==0.2.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
filesize-cli==1.0.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
groq==0.15.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
hashring==1.5.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
linecache2==1.0.0
memcache==0.12.0
multidict==6.1.0
openai==1.59.9
packaging==24.1
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.3
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-mock==1.12.1
rsa==4.9
setuptools==75.2.0
six==1.17.0
sniffio==1.3.1
tomli==2.0.2
tqdm==4.67.1
traceback2==1.4.0
typing_extensions==4.12.2
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
